Added toolbar and hotkeys for marks to slate
